% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package.R
\docType{data}
\name{clones_cana}
\alias{clones_cana}
\title{Médias de rendimento de dez clones de cana de açúcar avaliados em três épocas.}
\format{
Um \code{data.frame} com 30 observações:
\describe{
\item{\code{Clone}}{Fator de 10 níveis qualitativos, usado para
identificar os clones de cana de açúcar.}
\item{\code{Epoca}}{Fator de 3 níveis qualitativos, usado para
identificar a época  de avaliação.}
\item{\code{Taxa}}{Médias de rendimento}
}
}
\usage{
data(clones_cana)
}
\description{
O \code{data.frame} \code{clones_cana}, apresenta a médias de rendimento de
dez clones de cana de açúcar avaliados em três épocas.
}
\references{
Cruz, Cosme Damião & Regazzi, Adair José (1997). Modelos biométricos
aplicados ao melhoramento genético, Viçosa - UFV, 390p. (Tabela 8.6, p. 376)
}
\keyword{Repetibilidade}
\keyword{datasets}
