% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INeuICC.R
\docType{package}
\name{INeuICC}
\alias{INeuICC}
\title{Coeficiente de Correla\enc{çã}{ca}o Intra-Classe}
\description{
O pacote \code{INeuICC} re\enc{ú}{u}ne fun\enc{çõ}{co}es para a estimativa do
\emph{Coeficiente de Correla\enc{çã}{ca}o Intra-Classe}, baseado no Cap\enc{í}{i}tulo 8
do livro \strong{Modelos Biom\enc{é}{e}tricos Aplicados ao Melhoramento Gen\enc{é}{e}tico},
de Cosme Dami\enc{ã}{a}o Cruz e Adair Jos\enc{é}{e} Regazzi (1997).
}
\references{
Cruz, Cosme Dami\enc{ã}{a}o & Regazzi, Adair Jos\enc{é}{e} (1997). Modelos biom\enc{é}{e}tricos
aplicados ao melhoramento gen\enc{é}{e}tico, Vi\enc{ç}{c}osa - UFV, 390p. (Tabela 8.6, p. 376)
}
\author{
Ismael Neu (ismaelmmneu@hotmail.com)
}
