% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repetibilidade-aov-correl.R
\name{rpt_aov_fat1}
\alias{rpt_aov_fat1}
\title{Analise de Repetibilidade - ANOVA unifatorial}
\usage{
rpt_aov_fat1(dados, A, G, Y, R2 = 0.95)
}
\arguments{
\item{dados}{Data frame. Matriz de dados.}

\item{A}{String. Nome da variavel correspondendo a coluna da categoria Ambiente.}

\item{G}{String. Nome da variavel correspondendo a coluna da categoria Genotipo.}

\item{Y}{String. Nome da variavel correspondendo a coluna da categoria a
Variavel Resposta.}

\item{R2}{Numeric. Coeficiente de determinacao.}
}
\description{
Estima o coeficiente de repetibilidade, a analise de variancia.
Considera modelo com efeito de genotipo.
}
\details{
Modelo unifatorial utilizado: \eqn{Y_{ij} = m + G_i + e_{ij}}
}
\references{
Cruz, Cosme Damiao & Regazzi, Adair Jose (1997). Modelos biometricos
aplicados ao melhoramento genetico, Vicosa - UFV, 390p.
}
\author{
Ismael Neu (ismaelmmneu@hotmail.com)
}
