% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repetibilidade-componentes-principais.R
\name{rpt_cp_correl}
\alias{rpt_cp_correl}
\title{Analise de Repetibilidade - Componentes Principais - Correlacao}
\usage{
rpt_cp_correl(dados, A, G, Y, R2 = 0.95)
}
\arguments{
\item{dados}{Data frame. Matriz de dados.}

\item{A}{String. Nome da variavel correspondendo a coluna da categoria Ambiente.}

\item{G}{String. Nome da variavel correspondendo a coluna da categoria Genotipo.}

\item{Y}{String. Nome da variavel correspondendo a coluna da categoria a
Variavel Resposta.}

\item{R2}{Numeric. Coeficiente de determinacao.}
}
\description{
A funcao estima o coeficiente de repetibilidade, utilizando o metodo
dos componentes principais e a matriz de correlacoes.
}
\references{
Cruz, Cosme Damiao & Regazzi, Adair Jose (1997). Modelos biometricos
aplicados ao melhoramento genetico, Vicosa - UFV, 390p.
}
\author{
Ismael Neu (ismaelmmneu@hotmail.com)
}
